+++ src/app/components/space-transfers/transfers-table.js
@@ -55,8 +55,8 @@ const allColumnWidth = {
   userName: 200,
   destination: 200,
   scheduledAt: 150,
-  startedAt: 150,
-  finishedAt: 155,
+  startedAt: 130,
+  finishedAt: 130,
   processed: 110,
   replicated: 115,
   evicted: 95,
@@ -439,6 +439,7 @@ export default Component.extend(...mixins, {
     return this.createColumn('scheduledAt', {
       propertyName: 'scheduledAtReadable',
       className: 'hidden-xs',
+      component: 'cell-time',
     });
   }),
 
@@ -446,6 +447,7 @@ export default Component.extend(...mixins, {
     return this.createColumn('startedAt', {
       propertyName: 'startedAtReadable',
       className: 'hidden-xs',
+      component: 'cell-time',
     });
   }),
 
@@ -453,6 +455,7 @@ export default Component.extend(...mixins, {
     return this.createColumn('finishedAt', {
       propertyName: 'finishedAtReadable',
       className: 'hidden-xs',
+      component: 'cell-time',
     });
   }),