+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Latest changes
 
+* VFS-12278 Changed File ID column in the file browser to show the copy icon only on hover and allow copying the ID by clicking on the text
 * VFS-10341 Upgraded EmberJS to v3.16.10
 * VFS-12416 Fixed wrong "Eureka3D" mandatory literal to "EUreka3D" in EDM GUI
 * VFS-12277 Changed text "Basic" metadata to "xattrs" in modal tabs of item details